# docs/contracts/openapi/admin.openapi.yaml
openapi: 3.1.0
info:
  title: OnMaru Admin API
  version: 1.0.0-draft
  description: >-
    관리자 콘솔용 인증, 대시보드, 후기/신고 검수, 사용자 제재, 큐레이션,
    데이터 파이프라인 운영 계약. 모든 변경은 감사 로그와 멱등성 키를 사용한다.
servers:
  - url: /api/v1
...
paths:
...
  /admin/reviews/{reviewId}/moderation-actions:
    post:
      tags: [AdminReviews]
      operationId: moderateAdminReview
      summary: 후기 검수 상태 변경
      security: [{adminBearer: []}]
      parameters:
        - {name: reviewId, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModerationActionRequest'}
      responses:
        '200': {$ref: '#/components/responses/Review'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
...
  /admin/reviews:
    get:
      tags: [AdminReviews]
      operationId: listAdminReviews
      summary: 후기 목록 조회
      security: [{adminBearer: []}]
      parameters:
        - {$ref: '#/components/parameters/Status'}
        - {name: query, in: query, schema: {type: string, maxLength: 120}}
        - {name: reported, in: query, schema: {type: boolean}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: 후기 페이지
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReviewPage'}
        '401': {$ref: '#/components/responses/Error'}
  /admin/reports:
    get:
      tags: [AdminReports]
      operationId: listAdminReports
      summary: 신고 목록 조회
      security: [{adminBearer: []}]
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [PENDING, RESOLVED, REJECTED]}}
        - {name: reason, in: query, schema: {type: string}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: 신고 페이지
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportPage'}
        '401': {$ref: '#/components/responses/Error'}
...
  /admin/users/{memberId}/sanctions:
    get:
      tags: [AdminUsers]
      operationId: listAdminMemberSanctions
      summary: 회원 제재 이력 조회
      security: [{adminBearer: []}]
      parameters:
        - {name: memberId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: 제재 이력
          content:
            application/json:
              schema: {type: object, required: [schemaVersion, items], properties: {schemaVersion: {type: string}, items: {type: array, items: {$ref: '#/components/schemas/MemberSanction'}}}}
        '401': {$ref: '#/components/responses/Error'}
    post:
      tags: [AdminUsers]
      operationId: createAdminMemberSanction
      summary: 회원 제재 생성
      security: [{adminBearer: []}]
      parameters:
        - {name: memberId, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MemberSanctionRequest'}
      responses:
        '201':
          description: 제재 생성 완료
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MemberSanction'}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
  /admin/users/{memberId}/sanctions/{sanctionId}:
    delete:
      tags: [AdminUsers]
      operationId: revokeAdminMemberSanction
      summary: 회원 제재 해제
      security: [{adminBearer: []}]
      parameters:
        - {name: memberId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: sanctionId, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '204': {description: 제재 해제 완료}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminRefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-onmaru-admin-refresh
  parameters:
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 512}
    Status:
      name: status
      in: query
      schema: {type: string, enum: [PUBLISHED, HIDDEN, REMOVED]}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, format: uuid}
...
  headers:
    NoStore:
      schema: {type: string, const: no-store}
...
  responses:
    Error:
      description: 표준 API 오류
      headers:
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ApiError'}
    Review:
      description: 변경된 후기
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Review'}
  schemas:
...
    Review:
      type: object
      required: [id, status, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [PUBLISHED, HIDDEN, REMOVED]}
        content: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ReviewPage:
      type: object
      required: [items, hasNext]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Review'}}
        nextCursor: {type: string, nullable: true}
        hasNext: {type: boolean}
    ReportPage:
      type: object
      required: [items, hasNext]
      properties:
        items: {type: array, items: {type: object}}
        nextCursor: {type: string, nullable: true}
        hasNext: {type: boolean}
...
    MemberSanctionRequest:
      type: object
      required: [reason, startsAt]
      additionalProperties: false
      properties:
        reason: {type: string, minLength: 1, maxLength: 300}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time, nullable: true}
    MemberSanction:
      type: object
      required: [id, memberId, status, reason, startsAt, createdBy, createdAt]
      properties:
        id: {type: string, format: uuid}
        memberId: {type: string, format: uuid}
        status: {type: string, enum: [ACTIVE, REVOKED, EXPIRED]}
        reason: {type: string}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time, nullable: true}
        createdBy: {type: string, format: uuid}
        revokedBy: {type: string, format: uuid, nullable: true}
        revokedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
...
    ModerationActionRequest:
      type: object
      additionalProperties: false
      required: [nextStatus, reason]
      properties:
        nextStatus: {type: string, enum: [PUBLISHED, HIDDEN, REMOVED]}
        reason: {type: string, enum: [POLICY_VIOLATION, SPAM_CONFIRMED, FALSE_REPORT, OTHER]}
        note: {type: string, maxLength: 1000}
    ApiError:
      type: object
      required: [schemaVersion, code, message, requestId]
      properties:
        schemaVersion: {type: string}
        code: {type: string}
        message: {type: string}
        requestId: {type: string, format: uuid}
        details: {type: object}